[FIX] account: Fix reconciliation with very tiny rate with one foreign currency
Suppose a very tiny rate 1:10000.
l1 is expressed in company currency and has an amount of -10.
l2 is expressed in foreign currency and has 1000000 in company currency for 100 in foreign currency.

The reconciliation is made in foreign currency but there is nothing left to reconcile after converting
-10 to the foreign currency because -10 / 10000 = 0.0001 ~= 0.0
So no partial is created.

After this fix, the reconciliation fallbacks to a reconciliation in company currency if there is nothing
left to reconcile after applying the rate.

closes odoo/odoo#131756

Issue: 128234
X-original-commit: ba82fbe1b08108178a9fd7bade82b625c987e36e
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>